Labfiles/05-speech-tool/Python/speech-client/requirements.txt
python-dotenv
azure.identity
azure.ai.projects